# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PaperMan",
    "version": "0.1.0"
  },
  "paths": {
    "/api/catalog": {
      "get": {
        "summary": "Catalog",
        "operationId": "catalog",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Catalog"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/inboxes": {
      "get": {
        "summary": "Inboxes",
        "operationId": "inboxes",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Inbox"
                  },
                  "title": "Response Inboxes"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/inboxes/accounts": {
      "put": {
        "summary": "Register Inboxes",
        "operationId": "register_inboxes",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/AccountInbox"
                },
                "title": "Accounts"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Inbox"
                  },
                  "title": "Response Register Inboxes"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/inboxes/{inbox_id}/routing": {
      "put": {
        "summary": "Inbox Routing",
        "operationId": "inbox_routing",
        "parameters": [
          {
            "name": "inbox_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9_-]{1,80}$",
              "title": "Inbox Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InboxRouting"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Inbox"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccountInbox": {
        "properties": {
          "account_id": {
            "type": "string",
            "pattern": "^[a-zA-Z0-9_-]{1,80}$",
            "title": "Account Id"
          },
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "account_id",
          "name"
        ],
        "title": "AccountInbox"
      },
...
      "Catalog": {
        "properties": {
          "owners": {
            "items": {
              "$ref": "#/components/schemas/CatalogEntry"
            },
            "type": "array",
            "title": "Owners"
          },
          "tags": {
            "items": {
              "$ref": "#/components/schemas/CatalogEntry"
            },
            "type": "array",
            "title": "Tags"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "owners",
          "tags"
        ],
        "title": "Catalog"
      },
      "CatalogEntry": {
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^[a-zA-Z0-9_-]{1,80}$",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name"
          },
          "aliases": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Aliases"
          },
          "icon": {
            "type": "string",
            "enum": [
              "auto",
              "file",
              "receipt",
              "shield",
              "bank",
              "health",
              "utilities",
              "contract",
              "home",
              "car",
              "business",
              "education",
              "travel",
              "tax"
            ],
            "title": "Icon",
            "default": "auto"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "name",
          "aliases",
          "icon"
        ],
        "title": "CatalogEntry"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Inbox": {
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^[a-zA-Z0-9_-]{1,80}$",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name"
          },
          "account_id": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[a-zA-Z0-9_-]{1,80}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Account Id"
          },
          "routing_owner_ids": {
            "items": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9_-]{1,80}$"
            },
            "type": "array",
            "title": "Routing Owner Ids"
          },
          "time_format": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "24h",
                  "12h"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Format"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "name",
          "account_id",
          "routing_owner_ids",
          "time_format"
        ],
        "title": "Inbox"
      },
      "InboxRouting": {
        "properties": {
          "owner_ids": {
            "items": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9_-]{1,80}$"
            },
            "type": "array",
            "title": "Owner Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "owner_ids"
        ],
        "title": "InboxRouting"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}